# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Royal Ludo Arena API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matchmaking/join:
    post:
      operationId: joinMatchmaking
      tags: [matchmaking]
      summary: Join matchmaking queue for a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinMatchInput"
      responses:
        "200":
          description: Joined queue or matched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchStatus"
  /matchmaking/cancel:
    post:
      operationId: cancelMatchmaking
      tags: [matchmaking]
      summary: Cancel matchmaking and get refund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelMatchInput"
      responses:
        "200":
          description: Cancelled and refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchStatus"
  /matchmaking/status/{matchId}:
    get:
      operationId: getMatchStatus
      tags: [matchmaking]
      summary: Get match status
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchStatus"
...
  /withdrawals:
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal created and balance deducted immediately
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: Get withdrawals for current user
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/withdrawals/{id}/approve:
    patch:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{id}/reject:
    patch:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject a withdrawal and refund
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal rejected and refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
  /admin/notifications:
    post:
      operationId: sendNotification
      tags: [admin]
      summary: Send notification to all users or specific user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MatchStatus:
      type: object
      properties:
        matchId:
          type: ["string", "null"]
        status:
          type: string
        opponent:
          type: ["string", "null"]
        roomId:
          type: integer
        message:
          type: string
      required: [status, roomId, message]
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        method:
          type: string
        phone:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, userId, userName, amount, method, phone, status, createdAt]
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, title, message, isRead, createdAt]
...
    JoinMatchInput:
      type: object
      properties:
        roomId:
          type: integer
      required: [roomId]
    CancelMatchInput:
      type: object
      properties:
        matchId:
          type: string
      required: [matchId]
...
    WithdrawalInput:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
        phone:
          type: string
      required: [amount, method, phone]
...
    NotificationInput:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
      required: [title, message]
...